First step toward using header shims and macros instead of hipify_python
