+++ graphbolt/src/cache_policy.h
@@ -49,7 +49,11 @@ struct CacheKey {
         key_lower_32_bits_(key),
         position_in_cache_(position) {
     TORCH_CHECK(key == getKey());
+#ifndef GRAPHBOLT_USE_HIP
+    // TODO(#24): Removed assertion. Unknown why the size is important. Should
+    // pack bits.
     static_assert(sizeof(CacheKey) == 2 * sizeof(int64_t));
+#endif
   }
 
   CacheKey() = default;
@@ -133,7 +137,13 @@ struct CacheKey {
   int8_t freq_;
   // Negative values indicate writing while positive values indicate reading.
   // Access only through an std::atomic_ref instance atomically.
+#ifdef GRAPHBOLT_USE_HIP
+  // TODO(#24): rocm requires at least 4bytes for atomic_ref. int8_t doesn't
+  // work
+  int32_t reference_count_;
+#else
   int8_t reference_count_;
+#endif
   // Keys are restricted to be 48-bit unsigned integers.
   uint16_t key_higher_16_bits_;
   uint32_t key_lower_32_bits_;
