+++ graphbolt/src/cuda/cooperative_minibatching_utils.cu
@@ -23,10 +23,15 @@
 #include <thrust/transform.h>
 
 #ifdef __HIPCC__
+#include <hipsparse/hipsparse.h>
 #include <hipcub/hipcub.hpp>
+#include <cuco/static_map.cuh>
+#include <dgl/hip/cuda_to_hip.h>
+
 #else
 #include <cub/cub.cuh>
 #include <cuda/functional>
+
 #endif
 
 #include "../utils.h"
@@ -47,7 +52,14 @@ torch::Tensor RankAssignment(
         auto nodes_ptr = nodes.data_ptr<index_t>();
         THRUST_CALL(
             transform, nodes_ptr, nodes_ptr + nodes.numel(), part_ids_ptr,
-            ::cuda::proclaim_return_type<part_t>(
+            
+            #ifdef GRAPHBOLT_USE_HIP
+            ::proclaim_return_type
+            #else
+            ::cuda::proclaim_return_type
+            #endif
+            
+            <part_t>(
                 [rank = static_cast<uint32_t>(rank),
                  world_size = static_cast<uint32_t>(
                      world_size)] __device__(index_t id) -> part_t {
@@ -83,7 +95,7 @@ RankSortImpl(
                                               .dtype(offsets_dev.scalar_type())
                                               .pinned_memory(true));
         CUB_CALL(
-            DeviceFor::Bulk, num_batches * world_size + 1,
+            Bulk, num_batches * world_size + 1,
             [=, part_ids = part_ids_sorted.data_ptr<cuda::part_t>(),
              offsets = offsets.data_ptr<int64_t>()] __device__(int64_t i) {
               const auto batch_id = i / world_size;
