+++ graphbolt/src/cuda/unique_and_compact_impl.cu
@@ -22,7 +22,7 @@
 #include <thrust/functional.h>
 #include <thrust/gather.h>
 #include <thrust/logical.h>
-#ifdef __HIPCC__
+#ifdef GRAPHBOLT_USE_HIP
 #include <hipcub/hipcub.hpp>
 #else
 #include <cub/cub.cuh>
